app/requirements.txt
Flask==2.2
Flask-SQLAlchemy==2.5.1
psycopg2>2.0, <3.0
SQLAlchemy==1.4.25
pip==21.3.1
validate-docbr==1.10.0
python-dotenv
mercadopago
